requirements.txt
aiogram==3.6.0
aiosqlite==0.20.0
pydantic==2.7.0
python-dotenv==1.0.1
